MongoDB/Exercises/personal_finances/requirements.txt
flask-restplus==0.9.2
pymongo
Flask-PyMongo